+++ tests/x64_context_test.py
@@ -48,7 +48,7 @@ class X64ContextTests(jtu.JaxTestCase):
       {"testcase_name": "_jit={}".format(jit), "jit": jit}
       for jit in ["python", "cpp", None]))
   def test_make_array(self, jit):
-    func = _maybe_jit(jit, lambda: jnp.arange(10.0))
+    func = _maybe_jit(jit, lambda: jnp.arange(10.0, dtype=float))
     dtype_start = func().dtype
     with enable_x64():
       self.assertEqual(func().dtype, "float64")
@@ -66,7 +66,7 @@ class X64ContextTests(jtu.JaxTestCase):
     # The fact we defined a jitted function with a block with a different value
     # of `config.enable_x64` has no impact on the output.
     with enable_or_disable():
-      func = _maybe_jit(jit, lambda: jnp.arange(10.0))
+      func = _maybe_jit(jit, lambda: jnp.arange(10.0, dtype=float))
       func()
 
     expected_dtype = "float64" if config._read("jax_enable_x64") else "float32"
@@ -85,7 +85,7 @@ class X64ContextTests(jtu.JaxTestCase):
       self.skipTest("64-bit inverse not available on TPU")
     @partial(_maybe_jit, jit, static_argnums=1)
     def near_singular_inverse(key, N, eps):
-      X = random.uniform(key, (N, N))
+      X = random.uniform(key, (N, N), dtype=float)
       X = X.at[-1].mul(eps)
       return jnp.linalg.inv(X)
 
@@ -119,12 +119,12 @@ class X64ContextTests(jtu.JaxTestCase):
     def func_x32():
       with disable_x64():
         time.sleep(0.1)
-        return jnp.arange(10).dtype
+        return jnp.arange(10, dtype=int).dtype
 
     def func_x64():
       with enable_x64():
         time.sleep(0.1)
-        return jnp.arange(10).dtype
+        return jnp.arange(10, dtype=int).dtype
 
     with concurrent.futures.ThreadPoolExecutor() as executor:
       x32 = executor.submit(func_x32)
